+++ src/ClassicUO.Client/Configuration/Profile.cs
@@ -780,6 +780,8 @@ namespace ClassicUO.Configuration
 
         public double PaperdollScale { get; set => SetProperty(ref field, value); } = 1f;
 
+        public double StatusGumpScale { get; set => SetProperty(ref field, Math.Clamp(value, 0.5d, 3.0d)); } = 1f;
+
         public uint SOSGumpID { get; set => SetProperty(ref field, value); } = 1915258020;
 
         public bool ModernPaperdollAnchorEnabled { get; set => SetProperty(ref field, value); }